refactor: storybook templates
