+++ src/templates/storybook/vue/3/ts/Component.ts
@@ -1,23 +1,25 @@
-import type { ArgTypes, Meta, StoryFn } from '@storybook/vue3'
+import type { ArgTypes, Args, Meta, StoryFn } from '@storybook/vue3'
 
 import { toRefs } from 'vue'
 
 import ResourceName from 'resourcePath'
 
-const argTypes = {
+type Props = InstanceType<typeof ResourceName>
+
+const argTypes: Partial<ArgTypes<Props>> = {
   // Describe your args type (component properties) here to autodocs show properly informations
 }
 
-const meta = {
+const meta: Meta<Props> = {
   component: ResourceName,
-  title: 'Components/UI/Card',
+  title: 'Components/ResourceName',
   tags: ['autodocs'],
   argTypes
 }
 
 export default meta
 
-const Template = (args) => ({
+const Template: StoryFn<Props> = (args: Args) => ({
   components: { ResourceName },
   setup() {
     const reactiveArgs = toRefs(args)
@@ -25,13 +27,15 @@ const Template = (args) => ({
     return { args: reactiveArgs }
   },
   template: `
-    <ResourceName 
-      // pass here yours props, e.g: (remove this comment)
-
-      :id="args.id"
-    >
-        <!-- If your component receive slots, pass them here, if not, remove this comment -->
-    </ResourceName>
+    <div style="width: 100%; height: 100%; display: flex; justify-content: center; align-content: center; align-items: center;">
+      <ResourceName
+        :id="args.id"
+        
+        // pass here yours props, e.g: (remove this comment before run)
+      >
+          <!-- If your component receive slots, pass them here, if not, remove this comment -->
+      </ResourceName>
+    </div>
   `
 })
