+++ src/templates/storybook/vue/react/js/Component.js
@@ -1,0 +1,17 @@
+import { ResourceName } from 'resourcePath'
+
+const meta = {
+  component: ResourceName,
+  title: 'Components/ResourceName',
+  tags: ['autodocs'],
+  argTypes
+}
+
+export default meta
+
+export const Default = {
+  render: () => <ResourceName id="clingon-component" />,
+  args: {
+    // component props
+  }
+}
